Add Github release workflow
This workflow fully automates the process of releasing the package.
A release is triggered by a pull request that modifies the version in
`_version.py`.  The pull request must be submitted by an admin or a
maintainer.  The workflow then performs sanity checks on the version,
builds wheels on all supported platforms, uploads them to PyPI,
merges and tags the PR, and, finally, creates a Github release entry.