Add support for the std::bigint scalar type
`std::bigint` is an explicit variant of a numeric with zero fractional
scale.  It is an error to pass any value with fractional part as an
argument to a query accepting `bigint`.

I also took an opportunity to clean up and organize client-side argument
validation error reporting.  Invalid argument values now consistently
raise `InvalidArgumentValueError` which includes information on which
argument is problematic, as well as an excerpt from the value repr().