Rename AsyncConnection.close() to .aclose()
`close()` method in Python is a bit special as it makes an object
"closeable".  There are APIs that expect closeable objects to
shutdown when their `close` method is called. In the case of
`AsyncConnection`, old close() used to return a coroutine that
still needs to be awaited in order to do anything. Therefore, we
rename `close` to `aclose`, which also makes AsyncConnection symmetric
with async generators.

Closes #40.