Make sure aclose() waits for the server to close the connection
The point of the `aclose()` and the reason why it's a coroutine is to wait
for the server to close the connection.  This is known as graceful
disconnect, as opposed to a fast disconnect via `terminate()`.

Graceful disconnects are especially useful for cases when an agreement
on the state of the server after the disconnect is important, such as
when the script runs `DROP DATABASE` immediately after disconnecting.
If the disconnect is not negotiated properly, there is a possibility
that the server still thinks the connection is open and so a `DROP
DATABASE` operation would fail complaining that the database is still
being accessed.