Retry if _start() raises a retryable error
This issue was found in HA failover when the master Postgres is down, it
raises a BackendUnavailableError to start a transaction, but the
retrying_transaction() couldn't capture this retryable error from
start(), because start() is called in __enter__().

Now _start() is called in tx.execute() or query*() lazily.

Also removed start(), commit() and rollback() from
retrying_transaction().