Protect the first pool connection with a lock
Before this fix, the client allows multiple concurrent first
connections. This wasn't an issue because the _working_addr is always
the same among the concurrent first connections. However, it was not
clear which address shall win the race and be used for upcoming
connections when multiple addresses are used in the future. After this
fix, all concurrent first connection attempts will be serialized, the
address from the first successful connection will be used for following
connections. Concurrency will be resumed as soon as the address is
found, including the previously-serialized pending "first" connections.