Avoid an extra roundtrip in Execute on output tid mismatch
When only the output type descriptor passed by the client in `Execute`
does not match the server's, the server is still able to execute the
statement right away and provide the correct data descriptor alongside
data.  This saves an extra roundtrip in cases where the client knows the
input shape (e.g. there are no arguments, or the client is able to infer
the type descriptor from the types of values passed as arguments).