Ping first if conn is idle for too long
The blocking sockets don't have active connection_lost() events, if the
user is trying to execute a non-read-only query over a disconnected
connection, is_closed() cannot capture the disconnection and the retry
loop wouldn't work. In this case, we are sending a SYNC before the
query, so that we could capture the connection error and reconnect.

This is tailored to handle the server-side `session_idle_timeout`; for
regular network interruptions within the `session_idle_timeout`, we
don't want to sacrifice performance by inserting more frequent SYNCs
just to avoid connection errors on non-read-only queries.