Fix state of transaction start
This affects the case when compilation config is set on the client,
while a new command is compiled within a transaction. In this case, the
compiler will only use the state issued in the transaction start. Before
this fix, we didn't send state on transaction commands, so those config
was not in effect within transactions, even though each query did carry
the right config.