Prepare for 3.13 upgrade
New versions of cython have some regression in the handling of enums
that breaks us.  Fix it.  We have to do this separately from the
cython version upgrade because apparently we need to do the
edgedb-server cython upgrade at the same time. We'd like to just do
server first, but some test code on the server cimports edgedb-python,
and so we need to land the fix in edgedb-python before we can be ready
to upgrade the server.

This is of course a further argument in @mmastrac's favor about moving
towards a monorepo.