Add package alias with new name
Add a `gel` alias that does `import *` from `edgedb`.

Note that this will *not* provide any submodules.
Doing `from gel import errors` won't work.

But we also never documented anything internal, and it will all keep
working with an `edgedb` import, so nothing will break.

If we feel strongly about providing all the internal modules under the
`gel` name also, we have options, but as far as I can tell they are
all annoying in some way:
 - Generate a parallel tree of modules under `gel` that all do `import *`
   from an appropriate `edgedb` module. This is kind of ugly and also
   we need to regenerate when anything changes.
 - Something involving hooking the python import system. I think this
   would work but also I think involves modifying global python interpreter
   state when `gel` is imported, which is dodgy.

I'd probably go with option 1 if we care, but not caring seems fine?