Fix some ORM model generator issues.
Instead of hard errors issue warnings and attempt to skip over features
that cannot be implemented. A partial working set of models is better
than none and sometimes is what is needed.

Stop allowing multi properties in reflected models, they fundamentally
lack an identity and don't mesh with the Object models.